moved inner classes EventinfoListItem, EventCardItem and EventItemCardItem into own files
